refactor: add Types to AddAccountRepository
