+++ src/modules/account/data/protocols/add-account-repository.ts
@@ -1,3 +1,5 @@
+import { AddAccountModel } from '../../domain/models/add-account-model';
+
 export interface AddAccountRepository {
-  add(data: any): Promise<any>;
+  add(data: AddAccountModel.Params): Promise<void>;
 }
