+++ src/modules/account/domain/models/add-account-model.ts
@@ -1,0 +1,5 @@
+import { Account } from '../entities/Acount';
+
+export namespace AddAccountModel {
+  export type Params = ReturnType<Account['toJSON']>;
+}
