+++ src/modules/signup/domain/usecases/add-signup.ts
@@ -1,4 +1,4 @@
-import { SignupModel } from '../../data/models/add-signup';
+import { SignupModel } from '../models/add-signup';
 
 export interface AddSignup {
   add(params: SignupModel.Params): Promise<SignupModel.Result>;
